Forced cast int on filter return.

+++ plugins/woocommerce/includes/data-stores/class-wc-coupon-data-store-cpt.php
@@ -527,7 +527,7 @@ class WC_Coupon_Data_Store_CPT extends WC_Data_Store_WP implements WC_Coupon_Dat
 			$default_hold_time_minutes = 1;
 		}
 
-		return apply_filters( 'woocommerce_coupon_hold_minutes', $default_hold_time_minutes ) * 60;
+		return (int) apply_filters( 'woocommerce_coupon_hold_minutes', $default_hold_time_minutes ) * 60;
 	}
 
 	/**
